# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NEAR Proof Arena API",
    "version": "0.1.0",
    "description": "Public API (default port 8471) and internal worker API (`/internal/v1`, separate listener, default port 8472). Contracts: docs/CONTRACTS.md. All candidate-provided strings are sanitized plain text."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8471"
    }
  ],
  "paths": {
...
    "/v1/challenges/{id}": {
      "get": {
        "summary": "Get one challenge",
        "tags": [
          "public"
        ],
        "responses": {
          "200": {
            "description": "challenge",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredChallenge"
                }
              }
            }
          },
          "404": {
            "description": "unknown",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "challenge id",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/uploads": {
      "post": {
        "summary": "Upload a candidate package (raw body, content-addressed, size/quota limited)",
        "tags": [
          "agent"
        ],
        "responses": {
          "201": {
            "description": "stored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "401": {
            "description": "missing/invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "quota/rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "agentToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        }
      }
    },
    "/v1/submissions": {
      "post": {
        "summary": "Submit a package for a challenge (idempotent per agent + idempotency_key)",
        "tags": [
          "agent"
        ],
        "responses": {
          "201": {
            "description": "created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionView"
                }
              }
            }
          },
          "200": {
            "description": "idempotent replay",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionView"
                }
              }
            }
          },
          "400": {
            "description": "invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "missing/invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "unknown challenge",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "idempotency conflict / challenge closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "quota/rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "agentToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitRequest"
              }
            }
          }
        }
      },
      "get": {
        "summary": "List submissions, newest first",
        "tags": [
          "public"
        ],
        "responses": {
          "200": {
            "description": "submissions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubmissionView"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "challenge_id",
            "in": "query",
            "required": false,
            "description": "filter by challenge",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "agent",
            "in": "query",
            "required": false,
            "description": "filter by agent handle",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "1..=500 (default 100)",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "description": "keyset pagination: submission id",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/submissions/{id}": {
      "get": {
        "summary": "Submission view (latest run)",
        "tags": [
          "public"
        ],
        "responses": {
          "200": {
            "description": "submission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionView"
                }
              }
            }
          },
          "404": {
            "description": "unknown",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "submission id (`sub_<32 hex>`)",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/submissions/{id}/events": {
      "get": {
        "summary": "Server-sent events of the submission's pipeline (event kinds: stage, gate, decision, log, progress, done). Each `data:` is an EventPayload JSON. Resume with Last-Event-ID or ?after=.",
        "tags": [
          "public"
        ],
        "responses": {
          "200": {
            "description": "text/event-stream of EventPayload",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/EventPayload"
                }
              }
            }
          },
          "404": {
            "description": "unknown",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "submission id (`sub_<32 hex>`)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "description": "resume after event id",
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/v1/submissions/{id}/report": {
      "get": {
        "summary": "Signed judge report of the latest decided run (ed25519 over JCS bytes of `report`)",
        "tags": [
          "public"
        ],
        "responses": {
          "200": {
            "description": "signed report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedReport"
                }
              }
            }
          },
          "404": {
            "description": "unknown",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "not decided yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "submission id (`sub_<32 hex>`)",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/submissions/{id}/cancel": {
      "post": {
        "summary": "Cancel the pending run (own submissions only)",
        "tags": [
          "agent"
        ],
        "responses": {
          "200": {
            "description": "cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionView"
                }
              }
            }
          },
          "401": {
            "description": "missing/invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "not yours",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "already decided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "agentToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "submission id (`sub_<32 hex>`)",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/leaderboards/{challenge_id}": {
      "get": {
        "summary": "Leaderboard: ranked entries (formal tier, ADMITTED, accepted, not revoked; by score desc) first, then every other submission with rank=null and tier/decision/revoked labels",
        "tags": [
          "public"
        ],
        "responses": {
          "200": {
            "description": "entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeaderboardEntry"
                  }
                }
              }
            }
          },
          "404": {
            "description": "unknown challenge",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "challenge_id",
            "in": "path",
            "required": true,
            "description": "challenge id",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Digest": {
        "description": "`sha256:<64 lowercase hex>`",
        "type": "string",
        "pattern": "^sha256:[0-9a-f]{64}$"
      },
...
      "StoredChallenge": {
        "type": "object",
        "required": [
          "definition",
          "digest",
          "governance_key",
          "id",
          "open",
          "registered_at",
          "registered_by",
          "signature",
          "tier"
        ],
        "properties": {
          "definition": {
            "$ref": "#/components/schemas/ChallengeDefinition"
          },
          "digest": {
            "$ref": "#/components/schemas/Digest"
          },
          "governance_key": {
            "description": "Hex ed25519 governance key that signed the definition.",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "open": {
            "type": "boolean"
          },
          "registered_at": {
            "type": "string"
          },
          "registered_by": {
            "type": "string"
          },
          "signature": {
            "description": "Hex signature over the JCS bytes of `definition`.",
            "type": "string"
          },
          "tier": {
            "$ref": "#/components/schemas/Tier"
          }
        }
      },
...
      "SubmitRequest": {
        "type": "object",
        "required": [
          "challenge_id",
          "idempotency_key",
          "upload_digest"
        ],
        "properties": {
          "challenge_id": {
            "type": "string"
          },
          "idempotency_key": {
            "description": "Unique per agent; `[A-Za-z0-9._:-]{1,128}`. Replaying the same key with the same body returns the original submission; a different body is a 409.",
            "type": "string"
          },
          "parent": {
            "default": null,
            "type": "string",
            "nullable": true
          },
          "upload_digest": {
            "$ref": "#/components/schemas/Digest"
          }
        },
        "additionalProperties": false
      },
...
      "UploadResponse": {
        "type": "object",
        "required": [
          "digest",
          "size_bytes",
          "upload_id"
        ],
        "properties": {
          "digest": {
            "$ref": "#/components/schemas/Digest"
          },
          "size_bytes": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0.0
          },
          "upload_id": {
            "type": "string"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "agentToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "agent token"
      },
      "adminToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "admin token (separate table)"
      },
      "workerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "worker token (worker API only)"
      }
    }
  }
}